refactor(usePrepareAvailableTimeData): rename useEventTime
